rest/requirements.txt
Flask
Flask-RESTful

mysql-connector-python
flask_cors